9759 - refactoring the status logic
