+++ shared/src/business/dto/TrialSessionInfoDto.ts
@@ -1,28 +1,3 @@
-const { isEmpty, isEqual } = require('lodash');
-const { TRIAL_SESSION_SCOPE_TYPES } = require('../entities/EntityConstants');
-
-export const getTrialSessionStatus = ({ applicationContext, session }) => {
-  const { SESSION_STATUS_GROUPS } = applicationContext.getConstants();
-
-  const allCases = session.caseOrder || [];
-  const inactiveCases = allCases.filter(
-    sessionCase => sessionCase.removedFromTrial === true,
-  );
-
-  if (
-    session.isClosed ||
-    (!isEmpty(allCases) &&
-      isEqual(allCases, inactiveCases) &&
-      session.sessionScope !== TRIAL_SESSION_SCOPE_TYPES.standaloneRemote)
-  ) {
-    return SESSION_STATUS_GROUPS.closed;
-  } else if (session.isCalendared) {
-    return SESSION_STATUS_GROUPS.open;
-  } else {
-    return SESSION_STATUS_GROUPS.new;
-  }
-};
-
 export class TrialSessionInfoDto {
   public estimatedEndDate: string;
   public isCalendared: boolean;
@@ -43,10 +18,7 @@ export class TrialSessionInfoDto {
   public sessionScope: string;
   public sessionStatus: string;
 
-  constructor(
-    rawTrialSession: TTrialSessionData,
-    applicationContext: IApplicationContext,
-  ) {
+  constructor(rawTrialSession: TTrialSessionData) {
     this.estimatedEndDate = rawTrialSession.estimatedEndDate;
     this.isCalendared = rawTrialSession.isCalendared;
     this.isClosed = rawTrialSession.isClosed;
@@ -61,10 +33,6 @@ export class TrialSessionInfoDto {
     this.trialLocation = rawTrialSession.trialLocation;
     this.trialSessionId = rawTrialSession.trialSessionId;
     this.noticeIssuedDate = rawTrialSession.noticeIssuedDate;
-
-    this.sessionStatus = getTrialSessionStatus({
-      applicationContext,
-      session: rawTrialSession,
-    });
+    this.sessionStatus = rawTrialSession.sessionStatus;
   }
 }
